changed pathing + FS cleanup
