--- src/containers/App/modalStyles.js
@@ -1,13 +1,0 @@
-export default {
-  content: {
-    top: 400,
-    left: 400,
-    right: 400,
-    bottom: 400,
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    flexDirection: 'column',
-    border: 'none',
-  },
-}
