--- src/integrations/constants.ts
@@ -1,24 +1,0 @@
-export enum ETHEREUM_NETWORKS {
-  MAIN = 'MAIN',
-  MORDEN = 'MORDEN',
-  ROPSTEN = 'ROPSTEN',
-  RINKEBY = 'RINKEBY',
-  KOVAN = 'KOVAN',
-  UNKNOWN = 'UNKNOWN',
-}
-
-export enum WALLET_PROVIDER {
-  METAMASK = 'METAMASK',
-  PARITY = 'PARITY',
-  REMOTE = 'REMOTE',
-}
-
-export const EVENT_TYPES = {
-  CATEGORICAL: 'CATEGORICAL',
-  SCALAR: 'SCALAR',
-}
-
-export const ORACLE_TYPES = {
-  ULTIMATE: 'ULTIMATE',
-  CENTRALIZED: 'CENTRALIZED',
-}
