+++ src/integrations/providerWatcher.ts
@@ -1,10 +1,8 @@
 import { WalletProvider } from 'integrations/types'
-import { promisify } from 'api/utils'
-import { ETHEREUM_NETWORKS } from 'integrations/constants'
+import { promisify, shallowDifferent } from 'utils'
 import { Account, Balance } from 'types'
 import { getTime } from 'api'
-import { shallowDifferent } from 'utils/helpers'
-import { networkById } from 'integrations/initialize'
+import { ETHEREUM_NETWORKS, networkById } from 'globals'
 
 export const watcherLogger = ({ logType = 'log', status, info, updateState }: { logType: string, status: string, info: string, updateState: boolean }) =>
   console[logType](`
@@ -41,28 +39,28 @@ const providerInitAndWatcher = async (provider: WalletProvider, { updateMainAppS
     provider.state.timestamp = prevTime
 
     const [account, network, timestamp] = await Promise.all<Account, ETHEREUM_NETWORKS, number>([
-          getAccount(provider),
-          getNetwork(provider),
-          getTime(),
-        ]),
-        balance = account && await getBalance(provider, account),
-        available = provider.walletAvailable,
-        unlocked = !!(available && account),
-        newState = { account, network, balance, available, unlocked, timestamp }
+        getAccount(provider),
+        getNetwork(provider),
+        getTime(),
+      ]),
+      balance = account && await getBalance(provider, account),
+      available = provider.walletAvailable,
+      unlocked = !!(available && account),
+      newState = { account, network, balance, available, unlocked, timestamp }
 
       // if data changed
     if (shallowDifferent(provider.state, newState)) {
-        console.log('app state is different')
-        console.log('was: ', newState)
-        console.log('now: ', provider.state)
+      console.log('app state is different')
+      console.log('was: ', newState)
+      console.log('now: ', provider.state)
 
         // reset module timestamp with updated timestamp
-        prevTime = timestamp
+      prevTime = timestamp
         // dispatch action with updated provider state
-        updateProvider({ provider: provider.providerName, ...newState })
+      updateProvider({ provider: provider.providerName, ...newState })
         // check if initial load or wallet locked
 
-        if (!unlocked) {
+      if (!unlocked) {
           watcherLogger({
             logType: 'warn',
             status: 'WALLET LOCKED',
@@ -72,7 +70,7 @@ const providerInitAndWatcher = async (provider: WalletProvider, { updateMainAppS
           // if wallet locked, throw
           throw 'Wallet locked'
         }
-        else {
+      else {
           watcherLogger({
             logType: 'warn',
             status: 'CONNECTED + WALLET UNLOCKED',
@@ -81,7 +79,7 @@ const providerInitAndWatcher = async (provider: WalletProvider, { updateMainAppS
           })
           await updateMainAppState()
         }
-      }
+    }
   } catch (err) {
     console.warn(err)
       // if error
@@ -91,10 +89,10 @@ const providerInitAndWatcher = async (provider: WalletProvider, { updateMainAppS
 
     if (provider.walletAvailable) {
         // disable internal provider
-        provider.state.unlocked = false
+      provider.state.unlocked = false
         // and dispatch action with { available: false }
-        updateProvider({ provider })
-      }
+      updateProvider({ provider })
+    }
     throw err
   }
 }
