+++ src/middlewares/index.ts
@@ -1,9 +1,7 @@
 import CrashReporter from './CrashReporter'
-import NetworkChange from './NetworkChange'
 import NoScroll from './NoScroll'
 
 export {
   CrashReporter,
-  NetworkChange,
   NoScroll,
 }
