+++ src/utils/prototypes.ts
@@ -1,0 +1,4 @@
+// set last () => any on Array prototype
+Array.prototype.last = function getLast() {
+  return this[this.length - 1]
+}
